Add files via upload
v0.3
Naudojama Exception Handling
ir kodas išskaldytas į keletą failų, naudojamas header failas
